Update README.md
This repository contains a customized MATLAB code written along with the manuscript from Li et al. "Neurophysiological Connectomic Signatures of Consciousness during Propofol-Induced General Anesthesia". As this is adapted to a specific type of analysis, there are necessarily functions and parts of the codes that might not generalize so well to your data.